refactor(procedures): rename dm_restart_rco to dm_restart_r_co
